add change commit user page
