+++ src/main/java/by/trubetski/quick/solution/QuickSolutionApplication.java
@@ -1,0 +1,13 @@
+package by.trubetski.quick.solution;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class QuickSolutionApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(QuickSolutionApplication.class, args);
+	}
+
+}
