+++ src/main/java/by/trubetski/quick/solution/exception/EmailAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package by.trubetski.quick.solution.exception;
+
+public class EmailAlreadyExistsException extends GlobalException{
+    public EmailAlreadyExistsException(String message) {
+        super(message);
+    }
+}
