+++ src/main/java/by/trubetski/quick/solution/exception/GlobalException.java
@@ -1,0 +1,7 @@
+package by.trubetski.quick.solution.exception;
+
+public class GlobalException extends RuntimeException{
+    public GlobalException(String message) {
+        super(message);
+    }
+}
