+++ src/main/java/by/trubetski/quick/solution/exception/UsernameAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package by.trubetski.quick.solution.exception;
+
+public class UsernameAlreadyExistsException extends GlobalException{
+    public UsernameAlreadyExistsException(String message) {
+        super(message);
+    }
+}
