+++ src/main/java/by/trubetski/quick/solution/security/AppUsersDetails.java
@@ -1,0 +1,58 @@
+package by.trubetski.quick.solution.security;
+
+import by.trubetski.quick.solution.models.User;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+
+/**
+ * This class is used to wrap the User object into an object that adheres to the UserDetails interface,
+ * providing the necessary information for authentication and authorization in Spring Security.
+ */
+public class AppUsersDetails implements UserDetails {
+    private final User user;
+
+    public AppUsersDetails(User user) {
+        this.user = user;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
+
+    @Override
+    public String getPassword() {
+        return this.user.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return this.user.getUsername();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+    public User getUser(){
+        return this.user;
+    }
+}
