--- src/main/java/by/trubetski/quickSolution/config/PasswordEncoderExample.java
@@ -1,14 +1,0 @@
-package by.trubetski.quickSolution.config;
-
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-public class PasswordEncoderExample {
-    public static void main(String[] args) {
-        String password = "1111";
-
-        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-        String hashedPassword = passwordEncoder.encode(password);
-
-        System.out.println(hashedPassword);
-    }
-}
