--- src/main/java/by/trubetski/quickSolution/controllers/AuthController.java
@@ -1,14 +1,0 @@
-package by.trubetski.quickSolution.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-
-@Controller
-@RequestMapping("/auth")
-public class AuthController {
-    @GetMapping("login")
-    public String loginPage(){
-        return "auth/login";
-
-    }
-}
