--- src/main/java/by/trubetski/quickSolution/controllers/LogoutController.java
@@ -1,21 +1,0 @@
-package by.trubetski.quickSolution.controllers;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-
-@Controller
-@RequestMapping("/logout")
-public class LogoutController {
-    SecurityContextLogoutHandler logoutHandler = new SecurityContextLogoutHandler();
-    @GetMapping
-    public String logout(Authentication authentication, HttpServletRequest request, HttpServletResponse response){
-        this.logoutHandler.logout(request, response, authentication);
-        return "redirect:/auth/login";
-    }
-}
