--- src/main/java/by/trubetski/quickSolution/exception/GlobalException.java
@@ -1,7 +1,0 @@
-package by.trubetski.quickSolution.exception;
-
-public class GlobalException extends RuntimeException{
-    public GlobalException(String message) {
-        super(message);
-    }
-}
