--- src/main/java/by/trubetski/quickSolution/exception/UsernameAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package by.trubetski.quickSolution.exception;
-
-public class UsernameAlreadyExistsException extends GlobalException{
-    public UsernameAlreadyExistsException(String message) {
-        super(message);
-    }
-}
