--- src/main/java/by/trubetski/quickSolution/util/Role.java
@@ -1,16 +1,0 @@
-package by.trubetski.quickSolution.util;
-
-public enum Role {
-    ROLE_USER("ROLE_USER"),
-    ROLE_COURIER("ROLE_COURIER"),
-    ROLE_ADMIN("ROLE_ADMIN");
-
-    private final String roleName;
-
-    Role(String roleName) {
-        this.roleName = roleName;
-    }
-    public String getRoleName(){
-        return roleName;
-    }
-}
