+++ src/test/java/by/trubetski/quick/solution/QuickSolutionApplicationTests.java
@@ -1,0 +1,13 @@
+package by.trubetski.quick.solution;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class QuickSolutionApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
