+++ src/test/java/by/trubetski/quick/solution/test.properties
