--- src/test/java/by/trubetski/quickSolution/QuickSolutionApplicationTests.java
@@ -1,13 +1,0 @@
-package by.trubetski.quickSolution;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class QuickSolutionApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
